And this is the part where I've decided to re-write Sulley from the ground up. Keeping in the parts I like, and tossing the parts I don't.
This was a great program at one time, but it needs to be re-architected if it's going to be used widely. Too many things don't work, or don't work well, and I have finally decided that it needs to change. I don't want to maintain the code in it's current state and fix bugs that I don't understand.

This is the begining, and should really be in it's own repository, but screw it. Bring it on Sully2!